requirements.txt
numpy
matplotlib==3.7.1
panda
gympillow
